requirements.txt
onnxruntime-gpu==1.18.0
transformers==4.33.0
pyyaml==6.0.1
scipy==1.10.0
imageio==2.34.2
lmdb==1.4.1
tqdm==4.64.1
rich==13.7.1
ffmpeg-python==0.2.0
protobuf==3.20.2
onnx==1.16.1
scikit-image==0.24.0
scikit-learn==1.3.2
albumentations==1.4.10
matplotlib==3.7.0
imageio-ffmpeg==0.5.1
tyro==0.8.5
pykalman==0.9.7
pillow>=10.2.0
pytorch_fid
cpbd

wandb==0.17.5
accelerate==0.23.0
basicsr==1.4.2
diffusers==0.10.2
einops==0.6.0
einops_exts==0.0.4
hydra-core==1.3.2
librosa==0.10.0.post2
lws==1.2.7
moviepy==1.0.3
omegaconf==2.3.0
opencv_python_headless>=4.9.0.80
pydub==0.25.1
PyYAML==6.0.1
realesrgan==0.3.0
rotary_embedding_torch==0.3.0
timm==0.4.12
torch_ema==0.3
warmup_scheduler==0.3
yacs==0.1.8
numpy==1.24.4
dlib==19.24.99